requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
langchain-core
faiss-cpu
langchain-groq
langchain-community
langgraph
pydantic
pandas
tiktoken
unstructured
tavily-python
pypdf
python-docx
openpyxl
langsmith
sentence-transformers
langdetect
langchain_huggingface
langchain-google-genai
google-generativeai
msoffcrypto-tool
docx2txt
pdf2image
Pillow
